+++ sdk/network/azure-mgmt-network/azure/mgmt/network/v2020_11_01/aio/operations/_virtual_appliance_sites_operations.py
@@ -48,7 +48,7 @@ class VirtualApplianceSitesOperations:
         resource_group_name: str,
         network_virtual_appliance_name: str,
         site_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> None:
         cls = kwargs.pop('cls', None)  # type: ClsType[None]
         error_map = {
@@ -94,7 +94,7 @@ class VirtualApplianceSitesOperations:
         resource_group_name: str,
         network_virtual_appliance_name: str,
         site_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncLROPoller[None]:
         """Deletes the specified site from a Virtual Appliance.
 
@@ -106,8 +106,8 @@ class VirtualApplianceSitesOperations:
         :type site_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: Pass in True if you'd like the AsyncARMPolling polling method,
-         False for no polling, or your own initialized polling object for a personal polling strategy.
+        :keyword polling: By default, your polling method will be AsyncARMPolling.
+         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of AsyncLROPoller that returns either None or the result of cls(response)
@@ -163,7 +163,7 @@ class VirtualApplianceSitesOperations:
         resource_group_name: str,
         network_virtual_appliance_name: str,
         site_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> "_models.VirtualApplianceSite":
         """Gets the specified Virtual Appliance Site.
 
@@ -226,7 +226,7 @@ class VirtualApplianceSitesOperations:
         network_virtual_appliance_name: str,
         site_name: str,
         parameters: "_models.VirtualApplianceSite",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.VirtualApplianceSite":
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VirtualApplianceSite"]
         error_map = {
@@ -285,7 +285,7 @@ class VirtualApplianceSitesOperations:
         network_virtual_appliance_name: str,
         site_name: str,
         parameters: "_models.VirtualApplianceSite",
-        **kwargs
+        **kwargs: Any
     ) -> AsyncLROPoller["_models.VirtualApplianceSite"]:
         """Creates or updates the specified Network Virtual Appliance Site.
 
@@ -300,8 +300,8 @@ class VirtualApplianceSitesOperations:
         :type parameters: ~azure.mgmt.network.v2020_11_01.models.VirtualApplianceSite
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: Pass in True if you'd like the AsyncARMPolling polling method,
-         False for no polling, or your own initialized polling object for a personal polling strategy.
+        :keyword polling: By default, your polling method will be AsyncARMPolling.
+         Pass in False for this operation to not poll, or pass in your own initialized polling object for a personal polling strategy.
         :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of AsyncLROPoller that returns either VirtualApplianceSite or the result of cls(response)
@@ -360,7 +360,7 @@ class VirtualApplianceSitesOperations:
         self,
         resource_group_name: str,
         network_virtual_appliance_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.NetworkVirtualApplianceSiteListResult"]:
         """Lists all Network Virtual Appliance Sites in a Network Virtual Appliance resource.