+++ sdk/webpubsub/azure-messaging-webpubsubservice/CHANGELOG.md
@@ -1,9 +1,23 @@
 # Release History
 
-## 1.0.0b2 (Unreleased)
+## 1.0.0b3 (Unreleased)
 
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0b2 (2021-10-14)
+
+- Change api-version to `2021-10-01`  
+- Add operations to client
+- Support AAD
+- Support Api Management Proxy
 
 ## 1.0.0b1 (2021-04-27)
 
-Initial version
+- Initial version